refactor: Cleaned out the broken code, added comments and extracted CommunityBoundaries to a component.
